pyproject.toml
[project]
name = "whats-for-dinner"
version = "0.1.0"
description = ""
authors = [{ name = "Zaina Qasim", email = "zainaqsm@gmail.com" }]

requires-python = ">=3.12"
dependencies = [
    "haystack-ai==2.12.0",
    "fastapi>=0.115.0",
    "pillow>=10.4.0",
    "uvicorn[standard]>=0.27.0",
    "psycopg[binary]>=3.2.6",
    "pydantic-settings>=2.2.1",
    "pgvector-haystack>=5.5.0",
    "openai>=1.75.0",
    "python-multipart>=0.0.22",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["whats_for_dinner*"]


[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "pytest>=8.0.0",
    "anyio>=4.4.0",
    "httpx>=0.27.0",
    "ruff>=0.15.7",
    "basedpyright>=1.38.3",
]

[tool.pytest.ini_options]
testpaths = ["src"]
markers = [
    "integration: requires external services (Postgres, OpenAI)",
]
